requirements.txt
# Backend Framework
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# AI & Agents
langgraph>=0.0.10
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10

# Database & Data Processing
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
duckdb>=0.9.0
pandas>=2.0.0
numpy>=1.24.0

# ML & Monitoring
mlflow>=2.5.0
prometheus-client>=0.17.0
opentelemetry-api>=1.18.0
opentelemetry-sdk>=1.18.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
httpx>=0.24.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
